+++ documentation/code/NormalizedStackedBarChartDemo.tsx
@@ -30,16 +30,31 @@ export function NormalizedStackedBarChartDemo() {
               label: 'Direct',
               value: 200,
               formattedValue: '$200',
+              comparisonMetric: {
+                metric: '5',
+                trend: 'positive',
+                accessibilityLabel: 'Increase of',
+              },
             },
             {
               label: 'Facebook',
               value: 100,
               formattedValue: '$100',
+              comparisonMetric: {
+                metric: '5',
+                trend: 'negative',
+                accessibilityLabel: 'Decrease of',
+              },
             },
             {
               label: 'Twitter',
               value: 100,
               formattedValue: '$100',
+              comparisonMetric: {
+                metric: '5',
+                trend: 'neutral',
+                accessibilityLabel: 'Decrease of',
+              },
             },
             {
               label: 'Google',