+++ src/components/ComparisonMetric/components/index.ts
@@ -1,0 +1,2 @@
+export {UpChevron} from './UpChevron';
+export {DownChevron} from './DownChevron';