+++ src/components/ComparisonMetric/index.ts
@@ -1,0 +1,2 @@
+export {ComparisonMetric} from './ComparisonMetric';
+export {ComparisonMetricShape} from './types';