+++ src/components/NormalizedStackedBarChart/types.ts
@@ -1,3 +1,5 @@
+import {ComparisonMetricShape} from '../ComparisonMetric';
+
 export type Orientation = 'horizontal' | 'vertical';
 
 export type Size = 'small' | 'medium' | 'large';
@@ -6,4 +8,5 @@ export interface Data {
   formattedValue: string;
   value: number;
   label: string;
+  comparisonMetric?: ComparisonMetricShape;
 }